+++ src/main/java/frc/robot/constants/HangConstants.java
@@ -1,0 +1,9 @@
+package frc.robot.constants;
+
+public class Hang_Constants {
+    public static final int COMPRESSOR_PORT = 0; // Port for the compressor on the Pneumatics Hub
+    public static final int SOLENOID_1_CHANNEL = 1; // Channel for Solenoid 1
+    public static final int SOLENOID_2_CHANNEL = 2; // Channel for Solenoid 2
+    public static final int MIN_PRESSURE = 0; // Min pressure threshold to turn the compressor on
+    public static final int MAX_PRESSURE = 120; // Max pressure threshold to turn the compressor off
+}
